Fixed Critical Values using the Simulated Method

The title of each data set is the <Mother>_<Lugsail>_<alpha>_Master.

Bartlett_Mother_01_Master

Bartlett_Mother_025_Master

Bartlett_Mother_05_Master

Bartlett_Mother_10_Master

Bartlett_Zero_01_Master

Bartlett_Zero_025_Master

Bartlett_Zero_05_Master

Bartlett_Zero_10_Master

Bartlett_Over_01_Master

Bartlett_Over_025_Master

Bartlett_Over_05_Master

Bartlett_Over_10_Master

QS_Mother_01_Master

QS_Mother_025_Master

QS_Mother_05_Master

QS_Mother_10_Master

QS_Zero_01_Master

QS_Zero_025_Master

QS_Zero_05_Master

QS_Zero_10_Master

QS_Over_01_Master

QS_Over_025_Master

QS_Over_05_Master

QS_Over_10_Master

Parzen_Mother_01_Master

Parzen_Mother_025_Master

Parzen_Mother_05_Master

Parzen_Mother_10_Master

Parzen_Zero_01_Master

Parzen_Zero_025_Master

Parzen_Zero_05_Master

Parzen_Zero_10_Master

Parzen_Over_01_Master

Parzen_Over_025_Master

Parzen_Over_05_Master

Parzen_Over_10_Master

TH_Mother_01_Master

TH_Mother_025_Master

TH_Mother_05_Master

TH_Mother_10_Master

TH_Zero_01_Master

TH_Zero_025_Master

TH_Zero_05_Master

TH_Zero_10_Master

TH_Over_01_Master

TH_Over_025_Master

TH_Over_05_Master

TH_Over_10_Master

Format

A look up table containing the simulated robust critical values. The 'b' column contains the bandwidth, and subsequent columns are the dimensions

b: bandwidth of the test statistics. The proportion of autocovariances given a non-zero weight.

d: the dimension of the test statistic.

Details

The title of each data set is the <Mother>_<Lugsail>_<alpha>_Master. Supported mother kernels are Bartlett, Parzen, Tukey-Hanning (TH), and Quadratic Spectral (QS). Supported lugsail settings are mother, zero, and over. The available critical values are 0.01, 0.025, 0.05, and 0.10. Note that when b = 0 the robust critical value are equivalent to chi-square critical values.